Update FileType to be MultiDropzone specific
