+++ src/lib/components/multiDropzone/types.ts
@@ -1,17 +1,30 @@
 import { Accept } from "react-dropzone";
-import { FileType } from "../../types/file-type";
+
+export enum FileMimeTypes {
+  bmp = "image/bmp",
+  doc = "application/msword",
+  docx = "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
+  heic = "image/heic",
+  jpeg = "image/jpeg",
+  jpg = "image/jpg",
+  pdf = "application/pdf",
+  png = "image/png",
+  tif = "image/tiff",
+  tiff = "image/tiff",
+  webp = "image/webp",
+}
+
+export type FileType = keyof typeof FileMimeTypes;
 
 export type UploadStatus = 'UPLOADING' | 'COMPLETE' | 'ERROR';
 
 export const DOCUMENT_FILES: FileType[] = ['doc', 'docx', 'pdf'];
 export const IMAGE_FILES: FileType[] = ['heic', 'bmp', 'jpeg', 'jpg', 'png'];
 
-export const FILE_TYPES = [...DOCUMENT_FILES, ...IMAGE_FILES];
-
 export interface UploadedFile {
   id: string;
   name: string;
-  type: FileType | string;
+  type: FileType;
   previewUrl?: string;
   progress: number;
   error?: string;
